Isolando as responsabilidades para a camada de service, para que o código fique mais legível
