--- app/Http/Controllers/APILoginController.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-use Illuminate\Http\Request;
-
-class APILoginController extends Controller
-{
-    //Please add this method
-    public function login(Request $request) {
-        // get email and password from request
-        $credentials = $request->only('email', 'password');
-
-        // try to auth and get the token using api authentication
-        if (!$token = auth('api')->attempt($credentials)) {
-            // if the credentials are wrong we send an unauthorized error in json format
-            return response()->json(['error' => 'Unauthorized'], 401);
-        }
-        return response()->json([
-            'token' => $token,
-            'type' => 'bearer', // you can ommit this
-            'expires' => auth('api')->factory()->getTTL() * 60, // time to expiration
-
-        ]);
-    }
-}
