+++ app/Http/Controllers/UserController.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Service\UserService;
+use Illuminate\Http\Request;
+
+class UserController extends Controller
+{
+    private $service;
+
+    public function __construct(UserService $service)
+    {
+        $this->service = $service;
+    }
+
+    public function login(Request $request)
+    {
+        return $this->service->login($request);
+    }
+}
